+++ source/extensions/omni.flux.property_widget_builder.delegates/omni/flux/property_widget_builder/delegates/float_value/__init__.py
@@ -17,8 +17,10 @@
 
 __all__ = (
     "ColorField",
+    "FloatField",
     "FloatSliderField",
 )
 
 from .color import ColorField
+from .field import FloatField
 from .slider import FloatSliderField