+++ source/extensions/omni.flux.stage_manager.plugin.tree.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.6.1]
+### Changed
+- Removed unnecessary string-quoted type annotations
+
 ## [2.6.0]
 ### Added
 - Added NicknameField to the tree items